KBM-59
All of the previous devices that I implemented had a mouse component.
However, the ipv6buddy does NOT which causes a null pointer exception in
the pulling.  So I had to add a boolean to NOT check mouse.  This will
expand the compatiblity of this software for devices that do not have
mouse components.